requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
antlr4-python3-runtime==4.8
appdirs==1.4.4
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
autopage==0.5.0
backports.cached-property==1.0.1
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
einops==0.3.2
filelock==3.6.0
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
gast==0.5.3
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.0
h5py==3.6.0
huggingface-hub==0.5.1
HyperPyYAML==1.0.1
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
inaSpeechSegmenter==0.7.3
joblib==1.1.0
julius==0.2.6
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
librosa==0.9.1
llvmlite==0.38.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
networkx==2.8
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.1.2
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
pandas==1.4.2
pbr==5.9.0
Pillow==9.1.0
pooch==1.6.0
prettytable==3.3.0
primePy==1.3
protobuf==3.20.1
pyannote.algorithms==0.8
pyannote.audio @ https://github.com/pyannote/pyannote-audio/archive/develop.zip
pyannote.core==4.4
pyannote.database==4.1.3
pyannote.metrics==3.2
pyannote.parser==0.8
pyannote.pipeline==2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pyparsing==3.0.8
pyperclip==1.8.2
pyreadline3==3.4.1
Pyro4==4.82
PySocks==1.7.1
pytextgrid==0.1.4
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytorch-metric-learning==1.3.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
sentencepiece==0.1.96
serpent==1.40
shellingham==1.4.0
simplejson==3.17.6
singledispatchmethod==1.0
six==1.16.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
spectralcluster==0.2.4
speechbrain==0.5.11
SQLAlchemy==1.4.36
stevedore==3.5.0
sympy==1.10.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.5.4
tokenizers==0.12.1
torch==1.11.0
torch-audiomentations @ git+https://github.com/asteroid-team/torch-audiomentations@3304c7ae0004f8ca5f7b87a78d002969a5450144
torch-pitch-shift==1.2.2
torchaudio==0.11.0
torchmetrics==0.8.2
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
xarray==2022.3.0
xgboost==1.5.0
yarl==1.7.2
zipp==3.8.0
